Fix reply logic and move to contracts
