+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["packages/cheqd", "packages/common", "contracts/sdjwt-verifier", "packages/avida_test_utils", "packages/avida_example"]
+members = ["packages/cheqd", "packages/common", "contracts/sdjwt-verifier", "contracts/avida_example", "packages/avida_test_utils"]
 resolver = "2"
 
 [workspace.package]
