+++ contracts/avida_example/src/bin/schemas.rs
