+++ contracts/avida_example/src/constants.rs
@@ -7,5 +7,3 @@ pub const REGISTER_REQUIREMENT_REPLY_ID: RouteId = 1u64;
 pub const GIVE_ME_DRINK_ROUTE_ID: RouteId = 100u64;
 // Route id = 101
 pub const GIVE_ME_FOOD_ROUTE_ID: RouteId = 101u64;
-// Route id = 102
-pub const GIVE_ME_GASOLINE_ROUTE_ID: RouteId = 102u64;
