+++ contracts/avida_example/src/contract.rs
@@ -1,18 +1,20 @@
 use avida_common::traits::avida_verifier_trait::sv::AvidaVerifierTraitExecMsg;
 use avida_common::types::InputRoutesRequirements;
+
 use sylvia::cw_std::{to_json_binary, WasmMsg};
-use sylvia::cw_std::{entry_point, from_json, DepsMut, Env, Reply, Response, StdResult, SubMsg};
+use sylvia::cw_std::{from_json, Reply, Response, StdResult, SubMsg};
 use sylvia::types::ReplyCtx;
 use sylvia::{
     contract, entry_points, schemars,
     types::InstantiateCtx,
 };
+
 use cw_storage_plus::{Item, Map};
 use cw_utils::parse_reply_execute_data;
 
 use crate::error::ContractError;
-use crate::constants::{GIVE_ME_DRINK_ROUTE_ID, GIVE_ME_FOOD_ROUTE_ID, GIVE_ME_GASOLINE_ROUTE_ID, REGISTER_REQUIREMENT_REPLY_ID};
-use crate::msg::{ExecuteMsg, GiveMeSomeDrink, GiveMeSomeFood, GiveMeSomeGasoline, RegisterRequest, RegisterRequirement, VerifyRequest};
+use crate::constants::{GIVE_ME_DRINK_ROUTE_ID, GIVE_ME_FOOD_ROUTE_ID, REGISTER_REQUIREMENT_REPLY_ID};
+use crate::msg::{ExecuteMsg, GiveMeSomeDrink, GiveMeSomeFood, RegisterRequirement};
 
 
 pub struct RestaurantContract <'a>{
@@ -54,12 +56,6 @@ impl RestaurantContract <'_> {
                     requirements: requirements
                 }
             }
-            RegisterRequirement::Gasoline { requirements } => {
-                route_requirements = InputRoutesRequirements{
-                    route_id: GIVE_ME_GASOLINE_ROUTE_ID,
-                    requirements: requirements
-                }
-            }
         }
 
         let register_msg = AvidaVerifierTraitExecMsg::Register { 
@@ -81,12 +77,11 @@ impl RestaurantContract <'_> {
         Ok(Response::new().add_submessage(sub_msg))
     }
 
-    // // Ask for the portion
+    // Ask for the portion
     #[sv::msg(exec)]
     pub fn give_me_some_drink(&self, ctx: sylvia::types::ExecCtx, msg: GiveMeSomeDrink) -> StdResult<Response> {
         // 1. Save the transaction
         // 2. Send the request to verifier
-        // 3. Wait for the reply
         self.pending_transactions.save(ctx.deps.storage, GIVE_ME_DRINK_ROUTE_ID, &ExecuteMsg::GiveMeSomeDrink(msg.clone()))?;
         let verifier_contract = self.verifier.load(ctx.deps.storage)?;
         
@@ -109,45 +104,33 @@ impl RestaurantContract <'_> {
         )
     }
 
-    // // Ask for the portion
-    // #[sv::msg(exec)]
-    // pub fn give_me_some_food(&self, ctx: sylvia::types::ExecCtx, msg: GiveMeSomeFood) -> StdResult<Response> {
-    //     self.pending_transactions.save(ctx.deps.storage, GIVE_ME_FOOD_ROUTE_ID, &ExecuteMsg::GiveMeSomeFood(msg.clone()))?;
-    //     let verifier_contract = self.verifier.load(ctx.deps.storage)?;
-    //     let verify_request = SubMsg::reply_on_success( 
-    //         VerifyRequest{
-    //             presentation: msg.proof,
-    //             route_id: GIVE_ME_FOOD_ROUTE_ID,
-    //             app_addr: Some(ctx.info.sender.to_string())
-    //         }.into_cosmos_msg(verifier_contract)?,
-    //         GIVE_ME_FOOD_ROUTE_ID
-    //     );
+    // Ask for the portion
+    #[sv::msg(exec)]
+    pub fn give_me_some_food(&self, ctx: sylvia::types::ExecCtx, msg: GiveMeSomeFood) -> StdResult<Response> {
+        // 1. Save the transaction
+        // 2. Send the request to verifier
+        self.pending_transactions.save(ctx.deps.storage, GIVE_ME_FOOD_ROUTE_ID, &ExecuteMsg::GiveMeSomeFood(msg.clone()))?;
+        let verifier_contract = self.verifier.load(ctx.deps.storage)?;
+        let verify_request = AvidaVerifierTraitExecMsg::Verify {
+            presentation: msg.proof,
+            route_id: GIVE_ME_FOOD_ROUTE_ID,
+            app_addr: Some(ctx.env.contract.address.to_string()),
+        };
 
-    //     Ok(Response::new()
-    //         .add_submessage(verify_request)
-    //     )
+        let sub_msg = SubMsg::reply_on_success(
+            WasmMsg::Execute {
+                contract_addr: verifier_contract,
+                msg: to_json_binary(&verify_request)?,
+                funds: vec![],
+            },
+            GIVE_ME_FOOD_ROUTE_ID,
+        );
 
-    // }
+        Ok(Response::new()
+            .add_submessage(sub_msg)
+        )
 
-    // // Ask for the portion
-    // #[sv::msg(exec)]
-    // pub fn give_me_some_gasoline(&self, ctx: sylvia::types::ExecCtx, msg: GiveMeSomeGasoline) -> StdResult<Response> {
-    //     self.pending_transactions.save(ctx.deps.storage, GIVE_ME_GASOLINE_ROUTE_ID, &ExecuteMsg::GiveMeSomeGasoline(msg.clone()))?;
-    //     let verifier_contract = self.verifier.load(ctx.deps.storage)?;
-
-    //     let verify_request = SubMsg::reply_on_success(
-    //         VerifyRequest{
-    //             presentation: msg.proof,
-    //             route_id: GIVE_ME_GASOLINE_ROUTE_ID,
-    //             app_addr: Some(ctx.info.sender.to_string())
-    //         }.into_cosmos_msg(verifier_contract)?,
-    //         GIVE_ME_GASOLINE_ROUTE_ID,
-    //     );
-
-    //     Ok(Response::new()
-    //         .add_submessage(verify_request)
-    //     )
-    // }
+    }
 
     #[sv::msg(reply)]
     fn reply(&self, ctx: ReplyCtx, reply: Reply) -> Result<Response, ContractError> {
@@ -185,23 +168,6 @@ impl RestaurantContract <'_> {
                             return Ok(Response::new()
                                 .add_attribute("action", "give_me_some_food")
                                 .add_attribute("Food kind", msg.kind)
-                            )
-                        }
-                        _ => return Err(ContractError::VerificationProcessError)
-                    }
-                }
-                return Err(ContractError::VerificationProcessError)
-            }
-            GIVE_ME_GASOLINE_ROUTE_ID => {
-                let verification_result = parse_reply_execute_data(reply)?;
-                let verified: bool = from_json(&verification_result.data.unwrap())?;
-                let msg = self.pending_transactions.load(ctx.deps.storage, GIVE_ME_GASOLINE_ROUTE_ID)?;
-                if verified {
-                    match msg {
-                        ExecuteMsg::GiveMeSomeGasoline(msg) => {
-                            return Ok(Response::new()
-                                .add_attribute("action", "give_me_some_gasoline")
-                                .add_attribute("Amount of gasoline required", msg.amount)
                             )
                         }
                         _ => return Err(ContractError::VerificationProcessError)
