+++ contracts/avida_example/src/error.rs
@@ -13,9 +13,6 @@ pub enum ContractError {
     #[error("VerificationProcessError")]
     VerificationProcessError,
 
-    #[error("InternalError")]
-    InternalError,
-
     #[error("InvalidRouteId")]
     InvalidRouteId
 }
