+++ contracts/avida_example/src/lib.rs
