+++ contracts/avida_example/src/tests/keypair/test_ed25519_public.pem
