--- packages/avida_example/src/tests/fixtures.rs
@@ -1,39 +1,0 @@
-use jsonwebtoken::EncodingKey;
-use sd_jwt_rs::SDJWTIssuer;
-use serde_json::Value;
-use std::{fs, path::PathBuf};
-// Keys generation
-// ```sh
-// # for Ed25519
-// openssl genpkey -algorithm ED25519 -out private.pem
-// openssl pkey -in private.pem -pubout -out public.pem
-// ```
-
-pub fn issuer() -> SDJWTIssuer {
-    let mut key_path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    key_path = key_path.join("src/tests/fixtures/test_ed25519_private.pem");
-    println!("key_path: {:#?}", key_path);
-    let encoding_key_pem = fs::read(key_path).unwrap();
-    let encodingkey = EncodingKey::from_ed_pem(&encoding_key_pem).unwrap();
-    SDJWTIssuer::new(encodingkey, Some("EdDSA".to_string()))
-}
-
-pub fn issuer_jwk() -> josekit::jwk::Jwk {
-    let mut key_path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    key_path = key_path.join("src/tests/fixtures/test_ed25519_private.pem");
-    let encoding_key_pem = fs::read(key_path).unwrap();
-    let key_pair = josekit::jwk::alg::ed::EdKeyPair::from_pem(encoding_key_pem).unwrap();
-    println!("key_pair: {:#?}", key_pair);
-    key_pair.to_jwk_public_key()
-}
-
-pub fn claims(name: &str, age: u8, active: bool, joined_at: u16) -> Value {
-    serde_json::json!({
-        "exp": 1234567890,
-        "iss": "issuer",
-        "name": name,
-        "age": age,
-        "active": active,
-        "joined_at": joined_at
-    })
-}
